+++ Sources/Quiper/Constants.swift
@@ -9,6 +9,7 @@ struct Constants {
     static let WINDOW_CORNER_RADIUS: CGFloat = 10.0
     static let DRAGGABLE_AREA_HEIGHT: CGFloat = 32
     static let UI_PADDING: CGFloat = 5
+    static let SERVICE_REORDER_DRAG_THRESHOLD: CGFloat = 2
 
     static let STATUS_ITEM_OBSERVER_CONTEXT = 1